refactor: 임시 데이터 경로 변경
